+++ R/hierarchy_pull.R
@@ -47,24 +47,34 @@ hierarchy_pull <- function(ou_uid, username, password, baseurl = "https://final.
 hierarchy_clean <- function(df){
 
   #create header for each level of the org hierarchy
-  levels <- df$path %>%
-    stringr::str_count("/") %>%
-    max()
+    levels <- df$path %>%
+      stringr::str_count("/") %>%
+      max()
 
-  headers <- paste0("orglvl_", seq(1:levels))
+    headers <- paste0("orglvl_", seq(1:levels))
 
   #separate out path into each level of the org hierarchy (UIDs)
-  df <- df %>%
-    dplyr::mutate(path = stringr::str_remove(path, "^/")) %>%
-    tidyr::separate(path, headers, sep = "/", fill = "right") %>%
-    dplyr::select(-orglvl_1, -orglvl_2)
+    df <- df %>%
+      dplyr::mutate(path = stringr::str_remove(path, "^/")) %>%
+      tidyr::separate(path, headers, sep = "/", fill = "right") %>%
+      dplyr::select(-orglvl_1, -orglvl_2)
 
-  #convert level of the org hierarchy from UIDs to names
-  df_key <- dplyr::select(df, name, id)
+  #convert level names of the org hierarchy from UIDs to names
+    df_key <- dplyr::select(df, name, id)
+    df <- df %>%
+      dplyr::mutate_at(dplyr::vars(dplyr::starts_with("orglvl")),
+                       ~ plyr::mapvalues(., df_key$id, df_key$name, warn_missing = FALSE))
 
-  df <- df %>%
-    dplyr::mutate_at(dplyr::vars(dplyr::starts_with("orglvl")),
-                     ~ plyr::mapvalues(., df_key$id, df_key$name, warn_missing = FALSE))
+  #clean up coordinates, removing polygons and separating into lat-long
+    if(var_exists(df, "coordinates")) {
+      df <- df %>%
+        dplyr::mutate(coordinates = ifelse(stringr::str_detect(coordinates, "\\[{2,}"), NA, coordinates)) %>%
+        dplyr::mutate(coordinates = stringr::str_remove_all(coordinates, '\\[|]|"|[:space:]|\\t')) %>%
+        tidyr::separate(coordinates, c("longitude", "latitude"), sep = ",", extra = "drop") %>%
+        dplyr::mutate_at(dplyr::vars("longitude", "latitude"), as.double) %>%
+        dplyr::select(-longitude, dplyr::everything())
+    }
+
   return(df)
 }
 
@@ -101,21 +111,25 @@ hierarchy_rename <- function(df, country, username, password, baseurl = "https:/
 
     #clean up orgunits, keeping just OU, PSNU, Community and Facility
     df <- df %>%
-      dplyr::rename(orgunit = name,
-                    operatingunit = orglvl_3,
-                    snu1 = orglvl_4,
-                    orgunituid = id)
+      dplyr::rename_at(dplyr::vars(dplyr::matches("name|Organisation unit")), ~ "orgunit")
 
-    if(!!paste0("orglvl_", ou_country) %in% names(df)) {
+    if(var_exists(df, "id"))
+      df <- dplyr::rename(df, orgunituid = id)
+
+    if(ou_country == 3) {
       df <- df %>%
-        tibble::add_column(countryname = NA, .after = "operatingunit") %>%
-        dplyr::mutate(df, country = !!paste0("orglvl_", ou_country))
+        tibble::add_column(countryname = as.character(NA), .after = "orglvl_3") %>%
+        dplyr::mutate(countryname = orglvl_3)
+    } else {
+      df <- df %>%
+        tibble::add_column(countryname = as.character(NA), .after = "orglvl_3") %>%
+        dplyr::mutate(countryname = orglvl_4)
     }
 
     if("orglvl_4" %in% names(df)) {
       df <- df %>%
-        tibble::add_column(snu1 = NA, .after = "country") %>%
-        dplyr::mutate(df, snu1 = orglvl_4)
+        tibble::add_column(snu1 = NA, .after = "countryname") %>%
+        dplyr::mutate(snu1 = orglvl_4)
     }
 
     if(!!paste0("orglvl_", ou_psnu) %in% names(df))
@@ -132,9 +146,12 @@ hierarchy_rename <- function(df, country, username, password, baseurl = "https:/
     if(!!paste0("orglvl_", ou_fac) %in% names(df))
       df <- dplyr::rename(df, facility = !!paste0("orglvl_", ou_fac))
 
+    df <- dplyr::rename(df, operatingunit = orglvl_3)
+
     df <- df %>%
       dplyr::select(orgunit, orgunituid, dplyr::everything()) %>%
       dplyr::select(-dplyr::starts_with("orglvl_"))
+
 
     return(df)
   } else {
@@ -144,12 +161,33 @@ hierarchy_rename <- function(df, country, username, password, baseurl = "https:/
 }
 
 
+
+#' Extract country name from OU or country name
+#'
+#' @param df data frame created by `hierarchy_pull() %>% hierarchy_clean()`
+#'
+#' @export
+
+hierarchy_identify_ctry <- function(df){
+
+  #pull orglvl_3 which is out
+    country_name <- unique(df$orglvl_3)
+
+  #for regional missions, need to pull country name from orglvl_4
+    if(stringr::str_detect(country_name, "Region"))
+      country_name <- unique(df$orglvl_4) %>% setdiff(NA)
+
+  return(country_name)
+}
+
+
 #' Compile PEPFAR Hierarchy
 #'
 #' @param ou_uid UID for the country, recommend using `identify_ouuids()`
 #' @param username DATIM username
 #' @param password DATIM password, recommend using `mypwd()`
 #' @param baseurl API base url, default = https://final.datim.org/
+#' @param folderpath_output provide the full path to the folder for saving
 #'
 #' @export
 #'
@@ -160,7 +198,7 @@ hierarchy_rename <- function(df, country, username, password, baseurl = "https:/
 #' #pull hierarchy (paths are all UIDs)
 #'   df <- hierarchy_compile(ouuid, username = myuser, password = mypwd(myuser)) }
 
-hierarchy_compile <- function(ou_uid, username, password, baseurl = "https://final.datim.org/"){
+hierarchy_compile <- function(ou_uid, username, password, baseurl = "https://final.datim.org/", folderpath_output = NULL){
 
   print(ou_uid)
 
@@ -168,12 +206,12 @@ hierarchy_compile <- function(ou_uid, username, password, baseurl = "https://fin
 
   df <- hierarchy_clean(df)
 
-  ou_name <- unique(df$orglvl_3)
-  country_name <- dplyr::case_when(stringr::str_detect(ou_name, "Region") ~ unique(df$orglvl_4) %>% setdiff(NA),
-                                   TRUE ~ ou_name)
+  country_name <- hierarchy_identify_ctry(df)
 
   df <- purrr::map_dfr(.x = country_name,
                        .f = ~ hierarchy_rename(df, .x, username, password, baseurl))
+
+  hfr_export(df, folderpath_output, type = "mechanisms")
 
   return(df)
 }