+++ R/mech_pull.R
@@ -1,5 +1,9 @@
 #' Pull Partner/Mechanism Info from DATIM
 #'
+#' @param usaid_only specify if only USAID mechansism should be returned, default = TRUE
+#' @param ou_sel option to specify an operating unit, default = NULL
+#' @param folderpath_output provide the full path to the folder for saving
+#'
 #' @export
 #'
 #' @examples
@@ -7,7 +11,7 @@
 #' #pull mechanism/partner information
 #' df <- mech_pull() }
 
-mech_pull <- function(){
+mech_pull <- function(usaid_only = TRUE, ou_sel = NULL, folderpath_output = NULL){
 
   package_check("curl")
 
@@ -20,12 +24,32 @@ mech_pull <- function(){
     df <- readr::read_csv(sql_view_url,
                           col_types = readr::cols(.default = "c"))
 
-  #rename variables to match MSD and remove mech_code from mech_name
-  df <- df %>%
-    dplyr::select(mech_code = code,
-                  primepartner = partner,
-                  mech_name = mechanism) %>%
-    dplyr::mutate(mech_name = stringr::str_remove(mech_name_d, "0000[0|1] |[:digit:]+ - "))
+  #fitler for OU
+    if(!is.null(ou_sel))
+      df <- dplyr::filter(df, ou_sel %in% c(ou_sel))
+
+  #filter for USAID mechs if specified
+    if(usaid_only == TRUE)
+      df <- dplyr::filter(df, agency == "USAID")
+
+  #rename variables to match MSD
+    df <- df %>%
+      dplyr::select(operatingunit = ou,
+                    fundingagency = agency,
+                    mech_code = code,
+                    primepartner = partner,
+                    mech_name = mechanism)
+
+  #remove mech_code from mech_name
+    df <- df %>%
+        dplyr::mutate(mech_name = stringr::str_remove(mech_name, "0000[0|1] |[:digit:]+ - "))
+
+  #remove award information from mech_name
+    df <- df %>%
+      dplyr::mutate(mech_name = stringr::str_remove(mech_name, "^(720|AID|GHAG|U2GP).* - "))
+
+  #export
+    hfr_export(df, folderpath_output, type = "mechanisms")
 
   return(df)
 }